Model with combined observation and process error
This new model ditches the explicit state-space approach and instead directly models the observations only. Thus, the error term sigma_y includes process error and observation error. This may be fine given that there might not be information to separately estimate process vs. observation variance. The model works by constructing a full Y matrix which is partly observed, and defining the mean of the distribution with the "powered" AR1 specification. It seems like the parameters are reliably recovered. For datasets with less than 20 individuals, there do seem to be some issues with estimating the among-individual standard deviation in the expected value at timestep 1. Anyway, maybe this is helpful